src/cloud_functions/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform